requirements.txt
flask==3.0.2
flask-restful==0.3.10
flask-cors==4.0.0
gunicorn==21.2.0
faiss-cpu==1.10.0
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-ollama==0.2.3
protobuf==6.30.0
sentencepiece==0.2.0
numpy==1.26.4
addict==2.4.0
datasets==2.18.0
oss2==2.18.6
pillow==10.4.0
simplejson==3.19.3
sortedcontainers==2.4.0
modelscope==1.17.1
transformers==4.44.0
accelerate==1.4.0
fsspec==2024.2.0
torchvision==0.21.0
torchaudio==2.6.0
pytesseract==0.3.13
python-docx==1.1.2
markdownify==0.11.6
pypdf2==3.0.1
pdf2image==1.17.0
flask-restx==1.3.0
hjson
json5==0.12.0
lxml==5.3.2
opencv-python==4.11.0.86
poppler-utils==0.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
alembic==1.13.1
redis==6.1.0
email-validator==2.2.0
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.67.1
ipykernel==6.29.5
jupyter==1.1.1
matplotlib==3.10.3
PyJWT==2.10.1
bcrypt==4.0.1
pytest-shutil==1.7.0
pymupdf==1.23.5
jieba==0.42.1
minio==7.2.15
flask-caching==2.3.1
pgvector==0.4.1
cozepy==0.19.0
chardet==5.2.0